+++ sciris/sc_nested.py
@@ -10,6 +10,7 @@ Highlights:
 import re
 import itertools
 import pickle as pkl
+import collections as co
 from functools import reduce, partial
 import numpy as np
 import pandas as pd
@@ -17,7 +18,7 @@ from . import sc_utils as scu
 from . import sc_printing as scp
 
 # Define objects for which it doesn't make sense to descend further -- used here and sc.equal()
-_atomic_classes = (np.ndarray, pd.Series, pd.DataFrame, pd.core.indexes.base.Index) 
+atomic_classes = [np.ndarray, pd.Series, pd.DataFrame, pd.core.indexes.base.Index]
 
 
 ##############################################################################
@@ -219,7 +220,7 @@ def iternested(nesteddict, _previous=None):
     return output
 
 
-class IterObj(object):
+class IterObj:
     """
     Object iteration manager
     
@@ -273,10 +274,11 @@ class IterObj(object):
         print(all_data)
         
     | *New in version 3.1.2.*
+    | *New in version 3.1.5:* "norecurse" argument; better handling of atomic classes
     """    
-    def __init__(self, obj, func=None, inplace=False, copy=False, leaf=False, atomic='default', rootkey='root', verbose=False, 
-                _trace=None, _output=None, custom_type=None, custom_iter=None, custom_get=None, custom_set=None,
-                *args, **kwargs):
+    def __init__(self, obj, func=None, inplace=False, copy=False, leaf=False, recursion=0, 
+                 atomic='default', rootkey='root', verbose=False, _memo=None, _trace=None, _output=None, 
+                 custom_type=None, custom_iter=None, custom_get=None, custom_set=None, *args, **kwargs):
         from . import sc_odict as sco # To avoid circular import
         
         # Default argument
@@ -285,9 +287,11 @@ class IterObj(object):
         self.inplace   = inplace
         self.copy      = copy
         self.leaf      = leaf
+        self.recursion = recursion
         self.atomic    = atomic
         self.rootkey   = rootkey
         self.verbose   = verbose
+        self._memo     = _memo
         self._trace    = _trace
         self._output   = _output
         self.func_args = args
@@ -302,12 +306,22 @@ class IterObj(object):
         # Handle inputs
         if self.func is None: # Define the default function
             self.func = lambda obj: obj 
-        if self.atomic == 'default': # Handle objects to not descend into
-            self.atomic = _atomic_classes 
+            
+        atomiclist = scu.tolist(self.atomic)
+        if 'default' in atomiclist: # Handle objects to not descend into
+            atomiclist.remove('default')
+            atomiclist = atomic_classes + atomiclist
+        self.atomic = tuple(atomiclist)
+            
+        if self._memo is None:
+            self._memo = co.defaultdict(int)
+            self._memo[id(obj)] = 1 # Initialize the memo with a count of this object
+            
         if self._trace is None:
             self._trace = [] # Handle where we are in the object
             if inplace and copy: # Only need to copy once
                 self.obj = scu.dcp(obj)
+        
         if self._output is None: # Handle the output at the root level
             self._output = sco.objdict()
             if not inplace:
@@ -373,21 +387,26 @@ class IterObj(object):
         
         # Iterate over the object
         for key,subobj in self.iteritems():
-            trace = self._trace + [key]
-            newobj = subobj
-            subitertype = self.check_iter_type(subobj)
-            self.indent(f'Working on {trace}, leaf={self.leaf}, type={str(subitertype)}')
-            if not (self.leaf and subitertype):
-                newobj = self.func(subobj, *self.func_args, **self.func_kw)
-                if self.inplace:
-                    self.setitem(key, newobj)
-                else:
-                    self._output[tuple(trace)] = newobj
-            io = IterObj(self.getitem(key), self.func, inplace=self.inplace, leaf=self.leaf,  # Create a new instance
-                    atomic=self.atomic, verbose=self.verbose, _trace=trace, _output=self._output, 
-                    custom_type=self.custom_type, custom_iter=self.custom_iter, custom_get=self.custom_get, custom_set=self.custom_set,
-                    *self.func_args, **self.func_kw)
-            io.iterate() # Run recursively
+            newid = id(subobj)
+            if (newid in self._memo) and (self._memo[newid] > self.recursion): # If we've already parsed this object, don't parse it again
+                continue
+            else:
+                self._memo[newid] += 1
+                trace = self._trace + [key]
+                newobj = subobj
+                subitertype = self.check_iter_type(subobj)
+                self.indent(f'Working on {trace}, leaf={self.leaf}, type={str(subitertype)}')
+                if not (self.leaf and subitertype):
+                    newobj = self.func(subobj, *self.func_args, **self.func_kw)
+                    if self.inplace:
+                        self.setitem(key, newobj)
+                    else:
+                        self._output[tuple(trace)] = newobj
+                io = IterObj(self.getitem(key), self.func, inplace=self.inplace, leaf=self.leaf, recursion=self.recursion,  # Create a new instance
+                        atomic=self.atomic, verbose=self.verbose, _memo=self._memo, _trace=trace, _output=self._output, 
+                        custom_type=self.custom_type, custom_iter=self.custom_iter, custom_get=self.custom_get, custom_set=self.custom_set,
+                        *self.func_args, **self.func_kw)
+                io.iterate() # Run recursively
             
         if self.inplace:
             newobj = self.func(self.obj, *self.func_args, **self.func_kw) # Set at the root level
@@ -398,8 +417,8 @@ class IterObj(object):
             return self._output
         
 
-def iterobj(obj, func=None, inplace=False, copy=False, leaf=False, atomic='default', rootkey='root',
-            verbose=False, _trace=None, _output=None, *args, **kwargs):
+def iterobj(obj, func=None, inplace=False, copy=False, leaf=False, recursion=0, atomic='default', 
+            rootkey='root', verbose=False, _trace=None, _output=None, *args, **kwargs):
     """
     Iterate over an object and apply a function to each node (item with or without children).
     
@@ -421,6 +440,7 @@ def iterobj(obj, func=None, inplace=False, copy=False, leaf=False, atomic='defau
         inplace (bool): whether to modify the object in place (else, collate the output of the functions)
         copy (bool): if modifying an object in place, whether to make a copy first
         leaf (bool): whether to apply the function only to leaf nodes of the object
+        recursion (int): number of recursive steps to allow, i.e. parsing the same objects multiple times (default 0)
         atomic (list): a list of known classes to treat as atomic (do not descend into); if 'default', use defaults (e.g. ``np.array``, ``pd.DataFrame``)
         rootkey (str): the key to list as the root of the object (default ``'root'``)
         verbose (bool): whether to print progress.
@@ -457,9 +477,10 @@ def iterobj(obj, func=None, inplace=False, copy=False, leaf=False, atomic='defau
     | *New in version 3.1.0:* default ``func``, renamed "twigs_only" to "leaf", "atomic" argument
     | *New in version 3.1.2:* ``copy`` defaults to ``False``; refactored into class
     | *New in version 3.1.3:* "rootkey" argument
+    | *New in version 3.1.5:* "recursion" argument; better handling of atomic classes
     """
-    io = IterObj(obj=obj, func=func, inplace=inplace, copy=copy, leaf=leaf, atomic=atomic, rootkey=rootkey,
-                 verbose=verbose, _trace=_trace, _output=_output, *args, **kwargs) # Create the object
+    io = IterObj(obj=obj, func=func, inplace=inplace, copy=copy, leaf=leaf, recursion=recursion, atomic=atomic,
+                 rootkey=rootkey, verbose=verbose, _trace=_trace, _output=_output, *args, **kwargs) # Create the object
     out = io.iterate() # Iterate
     return out
 
@@ -766,7 +787,7 @@ def search(obj, query=_None, key=_None, value=_None, aslist=True, method='exact'
 class Equal(scp.prettyobj):
     
     # Define known special cases for equality checking
-    special_cases = (float,) + _atomic_classes
+    special_cases = tuple([float] + atomic_classes)
     valid_methods = [None, 'eq', 'pickle', 'json', 'str']
     
     
@@ -1112,8 +1133,9 @@ def equal(obj, obj2, *args, method=None, detailed=False, equal_nan=True, leaf=Fa
         - ``'pickle'``: converts the object to a binary pickle (most robust)
         - ``'json'``: converts the object to a JSON via ``jsonpickle`` (gives most detailed object structure, but can be lossy)
         - ``'str'``: converts the object to its string representation (least amount of detail)
+        - In addition, any custom function can be provided
     
-    By default, 'eq' is tried first, and if that raises an exception, 'pickle' is tried.
+    By default, 'eq' is tried first, and if that raises an exception, 'pickle' is tried (equivalent to ``method=['eq', 'pickle']``).
     
     Args:
         obj (any): the first object to compare